+++ changelog.md
@@ -3,6 +3,7 @@
 ## 0.1.1
 
 - add `fs/watchNodeFs.ts` for low level filesystem watching
+- expose `remove` and `ensureDir` in `fs/nodeFs.ts`
 
 ## 0.1.0